gnc_todo/pyproject.toml
[tool.poetry]
name = "gnc-todo"
version = "0.1.0"
description = "Uma api para gerenciamento de tarefas"
authors = ["Josué Luiz Barbosa e Silva <104951932+josuelb@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi-utils = "^0.7.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
testcontainers = "^4.7.2"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.111.1"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.31"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.5"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
fastapi-utils = "^0.7.0"
psycopg = "^3.2.1"
fastapi-redis-cache = "^0.2.5"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
